Refactor to use Long instead of Int everywhere.

+++ project-euler/src/main/scala/raymitchell/euler/problems/Problem12.scala
@@ -26,7 +26,7 @@ import raymitchell.euler.util.Math._
   */
 object Problem12 {
 
-  def solve(): Int =
+  def solve(): Long =
     TriangleNumberSequence
       .dropWhile(_.divisors.length <= 500)
       .head
